Improved backup system. Added config file

+++ restore-nugets.ps1
@@ -1,4 +1,7 @@
-$local:solutionDirectoryPath = "C:\Users\michal.korniak\source\repos\KRIP"
+function LoadConfigFile() {
+    $local:config = Get-Content -Path "config.json" | ConvertFrom-Json
+    return $config
+}
 
 function LogMessage($message) {
     $currentTime = Get-Date -format "dd-MMM-yyyy HH:mm:ss"
@@ -6,18 +9,19 @@ function LogMessage($message) {
 }
 
 function RestoreFileBackup([string]$filePath) {
-    $local:backupFilePath = ($filePath + '.backup')
-    Copy-Item -Path $backupFilePath -Destination $filePath
-    Remove-Item $backupFilePath
-    LogMessage ($filePath + ' backup file restored')
+    $local:backupFilePath = ($filePath + $config.BackupSufix)
+    if (Test-Path $backupFilePath -PathType Leaf) {
+        Copy-Item -Path $backupFilePath -Destination $filePath
+        Remove-Item $backupFilePath
+        LogMessage ($filePath + ' backup file restored')
+    }
 }
 
-
-
-$local:slnPath = (Get-ChildItem $SolutionDirectoryPath -Recurse *.sln).FullName
+$local:config = LoadConfigFile
+$local:slnPath = (Get-ChildItem $config.SolutionDirectoryPath -Recurse *.sln).FullName
 RestoreFileBackup $slnPath
 
-$local:projectsPaths = (Get-ChildItem $SolutionDirectoryPath -Recurse *.csproj).FullName
+$local:projectsPaths = (Get-ChildItem $config.SolutionDirectoryPath -Recurse *.csproj).FullName
 foreach ($projectPath in $projectsPaths) {
     RestoreFileBackup $projectPath
 }
